app/requirements.txt
Flask==1.0.0
psycopg2==2.7.4
Flask-SQLAlchemy==2.3.1
alembic==0.9.9